Consolidate E2E test helper functions into `helper.js`
E2E test helper functions were split between `helper.js` and `func.js`.
They've now all been moved to `helper.js`.
