+++ test/e2e/ethereum-on.spec.js
@@ -2,10 +2,8 @@ const assert = require('assert')
 const webdriver = require('selenium-webdriver')
 const { By, until } = webdriver
 const {
-  delay,
-} = require('./func')
-const {
   checkBrowserForConsoleErrors,
+  delay,
   findElement,
   openNewPage,
   verboseReportOnFailure,
