+++ test/e2e/helpers.js
@@ -3,16 +3,10 @@ const path = require('path')
 const mkdirp = require('mkdirp')
 const pify = require('pify')
 const assert = require('assert')
+const os = require('os')
+const { By, Builder, until } = require('selenium-webdriver')
+const { Command } = require('selenium-webdriver/lib/command')
 
-const {
-  delay,
-  getExtensionIdChrome,
-  getExtensionIdFirefox,
-  buildChromeWebDriver,
-  buildFirefoxWebdriver,
-  installWebExt,
-} = require('./func')
-const { until } = require('selenium-webdriver')
 const fetchMockResponses = require('./fetch-mocks.json')
 
 const tinyDelayMs = 200
@@ -23,6 +17,7 @@ module.exports = {
   assertElementNotPresent,
   checkBrowserForConsoleErrors,
   closeAllWindowHandlesExcept,
+  delay,
   findElement,
   findElements,
   openNewPage,
@@ -109,6 +104,56 @@ async function setupFetchMocking (driver) {
   await driver.executeScript(`(${fetchMocking})(${fetchMockResponsesJson})`)
 }
 
+function buildChromeWebDriver (extPath, opts = {}) {
+  const tmpProfile = fs.mkdtempSync(path.join(os.tmpdir(), 'mm-chrome-profile'))
+  const args = [
+    `load-extension=${extPath}`,
+    `user-data-dir=${tmpProfile}`,
+  ]
+  if (opts.responsive) {
+    args.push('--auto-open-devtools-for-tabs')
+  }
+  return new Builder()
+    .withCapabilities({
+      chromeOptions: {
+        args,
+        binary: process.env.SELENIUM_CHROME_BINARY,
+      },
+    })
+    .build()
+}
+
+function buildFirefoxWebdriver (opts = {}) {
+  const driver = new Builder().build()
+  if (opts.responsive) {
+    driver.manage().window().setSize(320, 600)
+  }
+  return driver
+}
+
+async function getExtensionIdChrome (driver) {
+  await driver.get('chrome://extensions')
+  const extensionId = await driver.executeScript('return document.querySelector("extensions-manager").shadowRoot.querySelector("extensions-item-list").shadowRoot.querySelector("extensions-item:nth-child(2)").getAttribute("id")')
+  return extensionId
+}
+
+async function getExtensionIdFirefox (driver) {
+  await driver.get('about:debugging#addons')
+  const extensionId = await driver.wait(until.elementLocated(By.xpath('//dl/div[contains(., \'Internal UUID\')]/dd')), 1000).getText()
+  return extensionId
+}
+
+async function installWebExt (driver, extension) {
+  const cmd = await new Command('moz-install-web-ext')
+    .setParameter('path', path.resolve(extension))
+    .setParameter('temporary', true)
+
+  await driver.getExecutor()
+    .defineCommand(cmd.getName(), 'POST', '/session/:sessionId/moz/addon/install')
+
+  return await driver.schedule(cmd, 'installWebExt(' + extension + ')')
+}
+
 async function checkBrowserForConsoleErrors (driver) {
   const ignoredLogTypes = ['WARNING']
   const ignoredErrorMessages = [
@@ -140,6 +185,10 @@ async function verboseReportOnFailure (driver, test) {
   await pify(fs.writeFile)(`${filepathBase}-screenshot.png`, screenshot, { encoding: 'base64' })
   const htmlSource = await driver.getPageSource()
   await pify(fs.writeFile)(`${filepathBase}-dom.html`, htmlSource)
+}
+
+function delay (time) {
+  return new Promise(resolve => setTimeout(resolve, time))
 }
 
 async function findElement (driver, by, timeout = 10000) {
