+++ test/e2e/incremental-security.spec.js
@@ -2,11 +2,9 @@ const assert = require('assert')
 const webdriver = require('selenium-webdriver')
 const { By, until } = webdriver
 const {
-  delay,
-} = require('./func')
-const {
   assertElementNotPresent,
   checkBrowserForConsoleErrors,
+  delay,
   findElement,
   findElements,
   openNewPage,
