+++ test/e2e/web3.spec.js
@@ -2,10 +2,8 @@ const assert = require('assert')
 const webdriver = require('selenium-webdriver')
 const { By } = webdriver
 const {
-  delay,
-} = require('./func')
-const {
   checkBrowserForConsoleErrors,
+  delay,
   findElement,
   findElements,
   openNewPage,
